third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-object-prototype-expected.txt

SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.

----- test getting values cross-frame using Object.prototype -----

PASS: Object.prototype.__lookupGetter__.call(targetFrame, 'myGetter'); should have thrown an exception, and did: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Object.prototype.__lookupSetter__.call(targetFrame, 'mySetter'); should have thrown an exception, and did: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Object.prototype.propertyIsEnumerable.call(targetFrame, 'myProp'); should have thrown an exception, and did: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.


--------
Frame: 'target'
--------

----- test that functions could not be injected cross-frame using Object.prototype -----

PASS: window.hasOwnProperty('injectedGetter') should be 'false' and is.
PASS: window.__lookupGetter__('injectedGetter') should be 'undefined' and is.
PASS: window.hasOwnProperty('injectedSetter') should be 'false' and is.
PASS: window.__lookupSetter__('injectedSetter') should be 'undefined' and is.
PASS: window.injectedGetter should be 'undefined' and is.
PASS: window.injectedSetter = 1 should be '1' and is.
